refactor(observe): own frame delivery and drop counts in the observer
The renderer loop counted failed channel sends into a public observer
field that Observer::observe then drained onto the next draft, splitting
one protocol across two modules. observe_into now detects, sends and
counts in one place; the carried-drop test moves beside it with the same
assertions.
