+++ crates/rimz/src/sidebar_pane/app/loop_state/tests/mod.rs
@@ -345,49 +345,3 @@ fn tripped_budget_ratchet_observes_the_day_spend_epoch_it_displays() {
     assert_eq!((usd, epoch), (3.0, Some(20)));
     assert_eq!(rig.state.ui.spend_ratchet.display(epoch, usd), 5.0);
 }
-
-#[test]
-fn failed_anomaly_send_preserves_carried_drop_count() {
-    // A zero-capacity channel fails every send, so this builds its own
-    // `LoopState` rather than taking the rig's buffered observe channel.
-    let ws = workspace();
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let runtime = RuntimePaths::under(ws.clone(), dir.path()).expect("runtime");
-    let (tx, _rx) = std::sync::mpsc::sync_channel(0);
-    let mut config = serve_config(&ws);
-    config.mux = MuxName::Tmux;
-    let socket_path = runtime.sidebar_socket_path(&config.instance_id);
-    let (_result_tx, result_rx) = std::sync::mpsc::channel();
-    let mut state = LoopState::new(
-        config,
-        runtime,
-        socket_path,
-        crate::diag::DiagSink::disabled(),
-        result_rx,
-        None,
-        tx,
-        PixelRenderCaps::default(),
-    );
-    let mut current = agent_snapshot(&ws);
-    let mut duplicate = current.worktree_groups[0].rows[0].clone();
-    duplicate.pane = duplicate.pane.map(|mut pane| {
-        pane.pane_id = crate::ids::PaneId::from_parts(crate::MuxName::Zellij, "terminal_10");
-        pane
-    });
-    current.worktree_groups[0].rows.push(duplicate);
-    current.worktree_groups[0].status_counts[0].count = 2;
-    state.current = current;
-    state.observer.dropped_msgs = 3;
-
-    state.observe_commit();
-
-    assert_eq!(
-        state.observer.dropped_msgs, 5,
-        "the failed anomaly send carries the prior drop count, then the failed roster send adds one"
-    );
-    state.observe_commit();
-    assert_eq!(
-        state.observer.dropped_msgs, 7,
-        "a consecutive full-channel commit keeps accumulating without losing the carried count or pending roster retry"
-    );
-}
